Refactor: separate test infra classes
