+++ src/test/java/com/chwihae/controller/index/IndexControllerTest.java
@@ -1,6 +1,6 @@
 package com.chwihae.controller.index;
 
-import com.chwihae.infra.AbstractMockMvcTest;
+import com.chwihae.infra.test.AbstractMockMvcTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
