+++ src/test/java/com/chwihae/controller/question/QuestionControllerTest.java
@@ -13,8 +13,8 @@ import com.chwihae.domain.vote.VoteEntity;
 import com.chwihae.dto.comment.request.QuestionCommentCreateRequest;
 import com.chwihae.dto.option.request.OptionCreateRequest;
 import com.chwihae.dto.question.request.QuestionCreateRequest;
-import com.chwihae.infra.AbstractMockMvcTest;
-import com.chwihae.infra.WithTestUser;
+import com.chwihae.infra.test.AbstractMockMvcTest;
+import com.chwihae.infra.support.WithTestUser;
 import com.chwihae.infra.fixture.UserEntityFixture;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
