+++ src/test/java/com/chwihae/domain/question/QuestionRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.chwihae.domain.question;
 
 import com.chwihae.domain.user.UserEntity;
-import com.chwihae.infra.AbstractIntegrationTest;
+import com.chwihae.infra.test.AbstractIntegrationTest;
 import com.chwihae.infra.fixture.UserEntityFixture;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
