+++ src/test/java/com/chwihae/service/comment/CommentServiceTest.java
@@ -10,7 +10,7 @@ import com.chwihae.domain.user.UserEntity;
 import com.chwihae.dto.comment.Comment;
 import com.chwihae.exception.CustomException;
 import com.chwihae.exception.CustomExceptionError;
-import com.chwihae.infra.AbstractIntegrationTest;
+import com.chwihae.infra.test.AbstractIntegrationTest;
 import com.chwihae.infra.fixture.UserEntityFixture;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
