+++ src/test/java/com/chwihae/service/user/UserServiceTest.java
@@ -3,7 +3,7 @@ package com.chwihae.service.user;
 import com.chwihae.domain.user.UserEntity;
 import com.chwihae.dto.user.UserContext;
 import com.chwihae.exception.CustomException;
-import com.chwihae.infra.AbstractIntegrationTest;
+import com.chwihae.infra.test.AbstractIntegrationTest;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
